Move identifiers to dedicated module
