+++ aries/misc/anoncreds_types/src/data_types/identifiers/mod.rs
@@ -1,0 +1,4 @@
+pub mod cred_def_id;
+pub mod issuer_id;
+pub mod rev_reg_def_id;
+pub mod schema_id;
