+++ aries/misc/anoncreds_types/src/data_types/identifiers/rev_reg_def_id.rs
@@ -1,0 +1,3 @@
+use crate::impl_anoncreds_object_identifier;
+
+impl_anoncreds_object_identifier!(RevocationRegistryDefinitionId);
