+++ aries/misc/anoncreds_types/src/data_types/ledger/mod.rs
@@ -1,5 +1,4 @@
 pub mod cred_def;
-pub mod issuer_id;
 pub mod rev_reg;
 pub mod rev_reg_def;
 pub mod rev_status_list;
