+++ aries/misc/anoncreds_types/src/data_types/messages/presentation.rs
@@ -1,9 +1,9 @@
 use std::collections::HashMap;
 
 use crate::cl::Proof;
-use crate::data_types::ledger::cred_def::CredentialDefinitionId;
-use crate::data_types::ledger::rev_reg_def::RevocationRegistryDefinitionId;
-use crate::data_types::ledger::schema::SchemaId;
+use crate::data_types::identifiers::cred_def_id::CredentialDefinitionId;
+use crate::data_types::identifiers::rev_reg_def_id::RevocationRegistryDefinitionId;
+use crate::data_types::identifiers::schema_id::SchemaId;
 use crate::error::ValidationError;
 use crate::utils::validation::Validatable;
