+++ aries/misc/anoncreds_types/src/data_types/mod.rs
@@ -1,4 +1,7 @@
+pub mod identifiers;
+#[cfg(feature = "ledger")]
 pub mod ledger;
+#[cfg(feature = "messages")]
 pub mod messages;
 
 pub mod macros;
